Improved IOInterface and Site classes
I added a method on the call to output the json file so that the "recording" boolean isn't included in the output. That matches his input file then.  I also added a set of reading ids  to the Site object and a .contains call to the add method because I had been able to put duplicates of the same reading into the object